requirements.txt
fastapi
uvicorn
Pillow
torch
transformers
